Include `use-cache` test suite in the Cache Components tests
We're splitting the test app routes between two root layouts, one for
(partially) static pages that does not wrap its children in a Suspense
boundary, and one for dynamic pages that adds a Suspense boundary above
`body`, which opts those pages into fully dynamic rendering.

One test is expected to fail for now, because it revealed a bug with
revalidation of prerendered route handlers. This will be fixed in a
follow-up PR.